Rename package to grpc_opts
